Tilføjet transactions til flertrinsmetoder i WishlistRepository. Puttet services i egne klasser.
